+++ src/finchlite/autoschedule/tensor_stats/dc_stats.py
@@ -12,7 +12,7 @@ from finchlite.algebra import Tensor
 from finchlite.algebra.algebra import FinchOperator
 from finchlite.finch_logic import Field
 
-from ._degree_scan import degree_count_scan
+from .util import degree_count_scan
 from .numeric_stats import NumericStats
 from .tensor_stats import BaseTensorStats, BaseTensorStatsFactory
